Observation: div(u) is exactly zero in all cells but one.
Partial explanation: the MMS problem has a nullspace of constant
pressures. When using LU, it doesn't pass the nullspace to PETSc,
but rather pins the pressure at a particular point (cell 0).
This pressure dof is then removed from the linear system, so
div(u) is unconstrained in that cell.

Unfortunately, when I try to pass the nullspace to PETSc
instead, it doesn't work; the output of MUMPS doesn't solve
the linear system (linear residual goes from 1 to 0.9).